+++ History.md
@@ -1,4 +1,6 @@
 - correcly write charge reason codes also for non-Xrechnung
+- correctly import additional referenced documents into invoice/corrected setting of attachments from jackson
+- corrected parseException structure
 
 2.15.1
 =======